CVPR 2018
Face-Age-Progression using GAN 
In contrast to the previous techniques that primarily operate on cropped facial areas (usually excluding foreheads), this paper emphasizes that synthesis of
the entire face is important since the parts of forehead and hair also significantly impact the perceived age. To achieve this and further enhance the aging details, their method leverages the intrinsic hierarchy of deep networks, and a discriminator of the pyramid architecture is designed to estimate high-level age-related clues in a fine-grained way.